+++ polymath/__init__.py
@@ -1,0 +1,1 @@
+from .library import Video  # noqa - required for compatibility with old database files
